backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
beautifulsoup4==4.14.3
requests==2.32.5